cam/requirements.txt
pyzbar
pandas
numpy
opencv-python
PyOpenGL
